change_varnames22.12.25

+++ ft_print_uint.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-int	ft_print_uint(unsigned long long uint)
+int	ft_print_uint(unsigned int uint)
 {
 	char *str;
 	int	len;
